requirements.txt
discord.py
pyfiglet
asyncio